+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - 实现搜索抽屉组件 `SearchDrawer` | Implement search drawer component `SearchDrawer`.
 - 支持 `i18n` | Support `i18n`.
 - 实现用户设置中的 `连接思源客户端` 组件 | Implement `Connect SiYuan Client` component in user settings.
+- 主界面显示服务状态 | Show service status in main interface.